+++ src/main/java/by/home/zoo/repository/BankAccountRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.repository.CrudRepository;
 /**
  * JPA repository for tour entity
  */
-public interface BankAccountRepository extends CrudRepository<BankAccount, Long> {
+public interface BankAccountRepository extends CrudRepository<BankAccount, Integer> {
 
 
 }
