+++ src/main/java/by/home/zoo/repository/SupplyStorageRepository.java
@@ -3,6 +3,6 @@ package by.home.zoo.repository;
 import by.home.zoo.entity.SupplyStorage;
 import org.springframework.data.repository.CrudRepository;
 
-public interface SupplyStorageRepository extends CrudRepository<SupplyStorage, Long> {
+public interface SupplyStorageRepository extends CrudRepository<SupplyStorage, Integer> {
 }
