+++ src/main/java/by/home/zoo/repository/ZooLogRepository.java
@@ -3,6 +3,6 @@ package by.home.zoo.repository;
 import by.home.zoo.entity.ZooLog;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ZooLogRepository extends CrudRepository<ZooLog, Long> {
+public interface ZooLogRepository extends CrudRepository<ZooLog, Integer> {
 }
