+++ src/main/java/by/home/zoo/repository/ZooRepository.java
@@ -3,6 +3,6 @@ package by.home.zoo.repository;
 import by.home.zoo.entity.Zoo;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ZooRepository extends CrudRepository<Zoo, Long> {
+public interface ZooRepository extends CrudRepository<Zoo, Integer> {
 }
